Move package-ci Go packages to internal package
