+++ package-ci/internal/pkg/templates/bridged-provider.config.yaml
