+++ package-ci/internal/pkg/templates/bridged-provider/.github/workflows/artifact-cleanup.yml
