+++ package-ci/internal/pkg/templates/bridged-provider/.github/workflows/command-dispatch.yml
