+++ package-ci/internal/pkg/templates/bridged-provider/.github/workflows/community-moderation.yml
