+++ package-ci/internal/pkg/templates/bridged-provider/.github/workflows/main.yml
