+++ package-ci/internal/pkg/templates/bridged-provider/.github/workflows/master.yml
