+++ package-ci/internal/pkg/templates/bridged-provider/.github/workflows/nightly-test.yml
