+++ package-ci/internal/pkg/templates/bridged-provider/.github/workflows/pull-request.yml
