+++ package-ci/internal/pkg/templates/bridged-provider/.github/workflows/release.yml
