+++ package-ci/internal/pkg/templates/bridged-provider/.github/workflows/resync-build.yml
