+++ package-ci/internal/pkg/templates/bridged-provider/.github/workflows/run-acceptance-tests.yml
