+++ package-ci/internal/pkg/templates/bridged-provider/.github/workflows/update-bridge.yml
