+++ package-ci/internal/pkg/templates/bridged-provider/.github/workflows/upgrade-bridge.yml
