+++ package-ci/internal/pkg/templates/bridged-provider/.github/workflows/upgrade-provider.yml
