+++ package-ci/internal/pkg/templates/bridged-provider/.golangci.yml
