+++ package-ci/internal/pkg/templates/bridged-provider/.goreleaser.prerelease.yml
