+++ package-ci/internal/pkg/templates/bridged-provider/.goreleaser.yml
