+++ package-ci/internal/pkg/templates/bridged-provider/.upgrade-config.yml
