+++ package-ci/internal/pkg/templates/bridged-provider/Makefile
