+++ package-ci/internal/pkg/templates/bridged-provider/scripts/upstream.sh
